requirements.txt
attrs==19.3.0
cycler==0.10.0
flake8==3.8.3
future==0.18.2
h5py==2.10.0
imageio==2.8.0
importlib-metadata==1.7.0
kiwisolver==1.2.0
matplotlib==3.2.2
mccabe==0.6.1
more-itertools==8.4.0
numpy==1.19.0
opencv-contrib-python==4.2.0.34
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
six==1.15.0
torch==1.5.1
torchvision==0.6.1
wcwidth==0.2.5
zipp==3.1.0
